tools/applause_detection/resources/requirements.txt
# was ==0.7.2
#librosa==0.7.2
librosa==0.8.0
# was >=1.21
numpy==1.22
# was ==0.9.2
numpydoc>=0.9.2
# was ==0.48
numba>=0.48
# was ==1.4.1
scipy>=1.4.1
# was ==0.22.1
scikit-learn>=0.22.1
# was ==0.2.0
ffmpeg-python>=0.2.0
tensorflow>=2.0.1
resampy==0.3.1
# me
requests>=2.25.0